Add an abstraction over StackGraph
