+++ src/components/shared/thread/HeightGraph.css
@@ -2,11 +2,11 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-.threadStackGraph {
+.threadHeightGraph {
   height: 25px;
 }
 
-.threadStackGraphCanvas {
+.threadHeightGraphCanvas {
   display: block;
   width: 100%;
   height: 25px;
